refactor: routes names to PascalCase to signify route group as entities, feat: saml2 certificates thinks
