+++ .gitignore
@@ -5,3 +5,5 @@ pulumi/.lambda
 .vscode
 !.vscode/tasks.json
 *.drawio.bkp
+certificates/*
+!certificates/*.sh
